[[ Mac64 ]] Update the externals API to be 64-bit clean
The window IDs were being passed as 32-bit values but this isn't
guaranteed to work on all platforms (in particular, it would break
for a 64-bit Windows port).

This is technically an ABI-incompatible change for 64-bit Linux
*but* the only affected external shipped by us is revbrowser
(which is new in 8 and we're still in DP) and there are no other
Linux externals that I know of that use this particular call.

+++ revbrowser/src/osxbrowser.mm
@@ -1273,12 +1273,12 @@ bool TAltBrowser::GetImage(void*& r_data, int& r_length)
 	return t_success;
 }
 
-int TAltBrowser::GetWindowId(void)
+uintptr_t TAltBrowser::GetWindowId(void)
 {
-	return (int)m_parent;
+	return uintptr_t(m_parent);
 }
 
-void TAltBrowser::SetWindowId(int p_new_id)
+void TAltBrowser::SetWindowId(uintptr_t p_new_id)
 {
 	NSWindow *t_window;
 	t_window = [NSApp windowWithWindowNumber: p_new_id];
